ORCID authorization

ORCID authorization

orcid_auth(scope = "/authenticate", reauth = FALSE,
  redirect_uri = getOption("rorcid.redirect_uri"))

Arguments

scope
(character) one or more scopes. default: "/authenticate"

reauth
(logical) Force re-authorization?

redirect_uri
(character) a redirect URI. optional

Value

a character string with the access token prefixed with "Bearer "

Details

There are two ways to authorise with rorcid:

Use a token as a result of a OAuth authentication process. The token is a alphanumeric UUID, e.g. dc0a6b6b-b4d4-4276-bc89-78c1e9ede56e. You can get this token by running orcid_auth(), then storing that key (the uuid alone, not the "Bearer " part) either as en environment variable in your .Renviron file in your home directory (with the name ORCID_TOKEN), or as an R option in your .Rprofile file (with the name orcid_token). See Startup for more information. Either an environment variable or R option work. If we don't find either we do the next option.

Interactively login with OAuth. This doesn't require any input on your part. We use a client id and client secret key to ping ORCID.org; at which point you log in with your username/password; then we get back a token (same as the above option). We don't know your username or password, only the token that we get back. We cache that token locally in a hidden file in whatever working directory you're in. If you delete that file, or run the code from a new working directory, then we re-authorize.

We recommend the former option. That is, get a token and store it as an environment variable.

If both options above fail, we proceed without using authentication. ORCID does not require authentication at this point, but may in the future - this prepares you for when that happens :)

ORCID OAuth Scopes

See https://members.orcid.org/api/orcid-scopes for more

Computing environments without browsers

One pitfall is when you are using rorcid on a server, and you're ssh'ed in, so that there's no way to open a browser to do the OAuth browser flow. Similarly for any other situation in which a browser can not be opened. In this case, run orcid_auth() on another machine in which you do have the ability to open a browser, then collect the info that's ouptput from orcid_auth() and store it as an environment variable (see above).

Examples

# NOT RUN {
x <- orcid_auth()
orcid_auth(reauth = TRUE)
#orcid_auth(scope = "/read-public", reauth = TRUE)
# }